Text-to-Image Sentiment analysis/requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.19.3
transformers==4.33.3
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
pyyaml==6.0.1
joblib==1.3.1
pytest==7.4.0
huggingface-hub==0.16.4
